Working on adding a git publish url as an export option.
The URL will determine where to publish the html files.
Typically, I like to keep my static websites in a separate
git repository and use github hosting (for example).

So the idea is to put the generated html into its own
git repository and push to the remote.

But we need to know where that remote is.

I would like it to be defined inside the index.html file.

Not sure exactly how to do this, but right now I am
able to read a custom option from the org file