Updated the macro replacement system.
Now, macros can take the form of ${pubme.XYZ},
and can be escaped with \${pubme.XYZ} to show them literally.

The only macro currently defined is ${pubme.BASE_DIR} which provides
a relative path to the base directory of the website that you want to
publish.  This allows us to put in proper paths for resources that
are not handled automatically by org-export (such as stylesheets)
in a consistent way.

So the stylesheet, for example, is specified as
${pubme.BASE_DIR}/pubme.css, and in the final pass this is
expanded to the proper relative directory, allowing all nested
files to use the same stylesheet code