Updated comments to test out a new deployment workflow.
Old workflow is still available, but the new workflow
allows github pages to be deployed from html by tracking in git.

This removes the need to have a source repository and deploy repository
and also enables tracking the output in git. If done from a private
github repository it means the full history is not visible.